Move different typing imports into compat module
