+++ pdm/models/versions.py
@@ -1,7 +1,7 @@
 import re
 from typing import Any, Tuple, Union, cast, overload
 
-from pdm._types import Literal
+from pdm.compat import Literal
 from pdm.exceptions import InvalidPyVersion
 
 VersionBit = Union[int, Literal["*"]]
